Modifie le type d'une probabilité d'une réaction en Double

+++ src/main/scala/model/Reaction.scala
@@ -12,7 +12,7 @@ case class Reaction(
   react: Solution,
   prod: Solution,
   k: Int,
-  a: Int
+  a: Double
 ) {
 
   /** Determine if the reaction can be apply on a given solution.
